Allow manual subscript on leaf nodes
Note, this is not perfect by any means... The subscript is not
factored into the calculation for the node width properly and the
combined text (value + subscript) is not properly centered in the
node.

There is room for a future refactor to calculate the combined text width
better and then layout the text while drawing in a way that properly
centers the text again.

Fixes: #2