Added phrase highlighing to the textarea
This will result in phrases with an unmatched closing bracket
to not be highlighted, but phrases with both brackets will be highlighted.

I created this feature for my professor as she mentioned
the brackts are hard to see for her (she has a visual impairment)

There have been slight edits to index.html and default.css,
as well as a new file highlight.js which contains
the logic behind the highlight feature.

Edited by: Andre Eisenbach <git@4ae.us>