Increase bottom arrow clearance slightlty
This fixes the arrows being clipped at the bottom with the example
sentence when "Align leaves" or "Align at bottom" are selected.

Ultimately, there might still be issues when the arrow can be clipped
or collide with other nodes.

The right fix here would be to make the arrow drawing "smarter" and
to better calculate the actual space requirement based on the curve
drawn.